16-preservation_DONE/pyproject.toml
[project]
name = "16-preservation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Aviksaikat", email = "saikickkarma@gmail.com" }
]
dependencies = [
    "eth-ape>=0.8.10",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/_16_preservation"]
